MAINT: move imag, round from ndarray to free functions
