+++ torch_np/tests/numpy_tests/lib/test_function_base.py
@@ -38,8 +38,7 @@ from torch_np import corrcoef, cov, i0, angle, sinc, diff, meshgrid, unique
 def get_mat(n):
     data = np.arange(n)
 #    data = np.add.outer(data, data)
-    from torch_np._helpers import _outer
-    data = _outer(data, data)
+    data = data[:, None] + data[None, :]
     return data
